Add files via upload
RaskVann wordle_solver - Changed bellerb's code to allow passing in previous guesses and the colour results into the bot. Changed word list to only consist of 5 letter words and moved old list to wordsOrig so iterations are faster. Changed authentication of guesses to also only allow guesses in the word list as that is part of the official wordle game rules.